third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-css-1-expected.txt

Verifies CSS pretty-printing functionality.


Running: testSimpleCSS
====== 8< ------
a {
    /* pre-comment */
    color /* after name */ : /* before value */ red /* post-comment */
}

------ >8 ======
Correct mapping for <pre-comment>
Correct mapping for <post-comment>

Running: testComplexCSS
====== 8< ------
@media screen {
    html {
        color: green;
        foo-property: bar-value
    }
}
}

body {
    background-color: black;
}

------ >8 ======
Correct mapping for <@media>
Correct mapping for <screen>
Correct mapping for <html>
Correct mapping for <color>
Correct mapping for <green>
Correct mapping for <foo-property>
Correct mapping for <bar-value>
Correct mapping for <body>
Correct mapping for <background>
Correct mapping for <black>

Running: testFormatInlinedStyles
====== 8< ------
<html>
    <body>
        <style>
            @-webkit-keyframes {
                from {
                    left: 0
                }

                to {
                    left: 100px;
                }
            }
        </style>
        <style>
            badbraces {
            }
        }

        @media screen {
            a.b {
                color: red;
                text-decoration: none
            }
        }
    </style>
</body>
</html>

------ >8 ======

Running: testNonZeroLineMapping
====== 8< ------
div {
    color: red;
}

------ >8 ======
Correct mapping for <div>
Correct mapping for <color>
Correct mapping for <red>

Running: testComplexSelector
====== 8< ------
a.b.c:hover,.d.e.f.g::before,h.i {
    color: red;
}

------ >8 ======
Correct mapping for <a>
Correct mapping for <.b>
Correct mapping for <.c>
Correct mapping for <.d>
Correct mapping for <.e>
Correct mapping for <.f>
Correct mapping for <.g>
Correct mapping for <h>
Correct mapping for <.i>
Correct mapping for <color>
Correct mapping for <red>
